More code tidyup.  An idea by Kenneth Albanowski -- we read the app/sort info at their natural time, which is backwards *after* we've read the entries backwards -- allows the code to be factored even better.  We probably do a minimum amount of duplication and carrying information around now -- it may even be comprehensible.
We also do some more (half-hearted) checking that the file being
read is a valid PRC/PDB file.  There may be more that we can do,
but the format has very little redundancy.

That's it for tidying up.  We can now do actual improvements:
for example, I don't think the Record data structure is very
convenient to use.  It should possibly have-a Datablock instead
of being is-a.

Probably the error reporting should be thought about too.