m68k-palmos-gdb in prc-tools 2.0 is unable to correctly print out array expressions such as "print a[3]".  I spent a while hunting for problems in our implementation of the remote-palmos protocol, but it turns out that it's not our fault at all.
It seems that the 4.18 release of GDB always gets this (and who
knows what else) wrong when cross-debugging from a little endian
host to a big endian target.  This is because people have done
explicit manual endianness hacking at a random layer in the protocol
stack.  Probably my particular number got turned around twice at
different layers of the stack or something.  This is what inevitably
happens when you do ad hoc manual endianness conversions in a
non-trivial program.  *Please* let's take this as a lesson.

This really sucks.  It's fixed in the current GDB CVS, and the hack
in this checkin papers over it for us in this release.

[Bug report from Niklas Frykholm.]