In the absence of any other symbols, use MacsBUG symbols from the text (eg the ROM) for a function name (eg in a stack backtrace).
Of course, this isn't the right way to plug the functionality
into stack.c.  We should at least have a TARGET_LAST_CHANCE_HOOK
sort of a macro, or even better maybe a vector in target_ops.
But we'll sort that out when we submit this patch back.

(Wow!  This one's been on my to do list since 1998!)