Start to get GLibs working again.  Here's what's involved:
* (Makefile stuff)  Do a form of multilibbing so that we can have
  one set of startup code using %a5 for globals (for applications)
  and another using %a4 (for shared objects).  This may not be the
  final form of multilibbing used: I've been meaning to check
  exactly what GCC's multilibbing looks like (eg in the names of
  the installed .a files) and whether we should be using the same
  system.

* (crt files, especially palmos_Glib.c)  Build differently
  depending on whether __OWNGP__ is set, which in turn is set by
  the spec file when -mown-gp is used, which is done for one of
  the sets by the magical Makefile stuff.

I did this work a while back, and checking it in has been delayed
while I wanted to do some testing.  I'm checking it in now so that
some other people can do some testing (thanks David!).  This is
likely not to be the final form of this stuff, but it is certainly
the beginning.