Add a new target specific option, -mexperimental-return-reg-d0, which breaks the ABI by generating functions which return pointers in %d0 (like scalars) instead of %a0.  Note that return types such as double will still be returned in %d0/%d1.
Note that code compiled with this option cannot call Palm OS API
functions or library functions which do comply with the ABI.  So
this option should not in general be used!

Some folks have various special needs and would like to experiment
with this calling convention.  This option is not supported and
may disappear in the future.  But experience with it will help
determine how best to satisfy these needs.  [Mark Lummus]