Undo the twisty turny -palmosX spec file stuff that I spent such a lot of time on eighteen months ago.  That method always wanted just another couple of features in the specs parser to work properly, and, while that was tempting :-), I was always conscious that the GCC maintainers' constant refrain when asked about such things was always "just tell the users to wrap a script around GCC".  (Gak!)
So m68k-palmos-sdkfind is a wrapper to be used instead of GCC :-).
It's not so horrible.  The real m68k-palmos-gcc now knows nothing
about /usr/local/palmdev and makes no attempt to supply -I options
to see in there.  Instead, when sdkfind is installed in the right
way, it will rewrite any -palmosX options into mountains of -isystem
and -L options, and then call down to the real m68k-palmos-gcc.  It
also acts as a filter, so "m68k-palmos-gcc foo.c `m68k-palmos-sdkfind`"
will do the right thing.

With this patch it gets built and installed, but not plugged in.
We'll probably want to rename the installed m68k-palmos-gcc as
real-m68k-palmos-gcc, and install sdkfind as m68k-palmos-gcc.
Or something.  It's all pretty tacky.  Perhaps there are other
solutions that balance confusion and tackiness better...

The handspring.palmos.diff patch is no longer necessary, since
sdkfind will automatically look inside an include/Handspring
directory if there is one.