Complete overhaul of error reporting routines.  It turns out that the experiment with having implicit state in the filename and lineno globals was a failure.  (Duh!)  Surprisingly enough, sometimes calls of einfo() would forget to update filename, leading to misleading error messages referring to the wrong files.
Now the filename parameters etc are explicit.  Some error handling had
to be rearranged a little so that those parameters were passed down to
where they were needed.

Also changed all the error messages to quote things 'like this' instead
of `like this'.  We're now Markus Kuhn compliant :-).  (See
http://www.cl.cam.ac.uk/~mgk25/ucs/quotes.html )

progname and progversion are next on the hit list.