Aargh.  This is horrible.  We want to use the section "[eb]hook" attribute on variables (eg void *foo __attribute__ (section ("bhook")) = myhookfunc;) to add entries to the [eb]hook sections.  (This is more reliable in the face of the optimizer than trying to do it via inline asm.  And doing it in C is Just Plain Nicer.)
But doing that marks [eb]hook as a *data* section.  And we try to emit .reloc
stuff for it.  This is a error -- see the comment in the patch.  Fortunately,
some day soon we'll upgrade to 2.11 with its --emit-relocs, and hopefully
leave --embedded-relocs behind forever.  (The person who added the broken
version of m68k --embedded-relocs to 2.11 can use it, but we won't need to.
Quite ironic, really.)

For now, we just hack around it.