Put these headers in {exec_prefix}/{target_alias}/include instead of {palmdev_prefix}/include, since policy is now not to write to {palmdev_prefix} at install time.
This is suboptimal because these headers are supposed to be target
independent, so it's silly to duplicate them across targets.  (That's
half the reason PalmDev exists, dammit.)  Eventually maybe we'll have
a {datadir}/prc-tools/include (which would be pretty ironic :-)), and
maybe even symlink it into PalmDev, but we should wait and see what
Palm does with ARM headers first.