Total overhaul of the top-level build mechanism.  The GCC and target builds now have fingers reaching over into the previous build directories (for binutils and GCC) so that they can use the not-yet-installed utilities from those directories.
This means we can build with "make all" and install with "make
install", instead of the cursed interleaved build binutils/install it/
build gcc/install it/build target/install it process.  (The proximate
need for this was the RPM build root hack, which didn't mix at all
well with this interleaving.)

doc/Makefile.in: Bring back the "info" target, so that we can still
say "make all info" in the top-level.  Oops.  We won't be bringing
back install-info as is: the top-level "install" target still has
some work wanting to be done on it.

This change breaks --with-build-sdk; you can currently only use the
bootstrap headers and will get many "no such option -nopalmos" errors.
This is because install-wrappers (delenda est install-wrappers) isn't
fingered; it still only happens at installation time.  The correct
(and rather farreaching!) response is to get rid of sdkfind and
install-wrappers, which were always nasty hacks.

Instead, we're going back to setting the palmdev paths in GCC's specs.
Rather than try to do the impossible by making the specs at build time,
we'll supply a new tool that users can use to generate a specs file
customised for their installation.  Duh.  It all seems so simple now!

This is better for the build process, faster for the users (no more
running sdkfind on every GCC invocation), and generally Less Tacky.