Initial version of updated binutils patches applying to binutils 2.11.2.  This initial checkin is just the configury to enable the targetting of m68k-palmos.
Previously we had config.sub hacks to rewrite m68k-palmos as
m68k-palmos-coff.  Now with this configury we can have m68k-palmos
as a real target OS (and we have a config.sub hack to rewrite
m68k-palmos-coff as m68k-palmos).

The subdirectory configury patches are in their final form.  The
config.sub patch is a quick hack to make it the same as what is/
will be in mainline config.sub.