Add the R_RELENDWORD reloc and its handling to bfd.  This is what we've used to date to handle the fact that %a5 points to the *end* of the global data, not the beginning.
This is a slightly tidier version than the older 2.9.1 one: it's
recoded so as to touch only coff-m68k.c, not cofflink.c as well.

We may change to some other way of handling the %a5 problem in the
future.  But we'll need to keep this patch around for a while to be
able to handle relocs lying around in old libraries.