Palmdev-prep's main job is to generate a GCC spec file listing the available SDKs and other material in PalmDev.  It is a replacement for sdkfind, which has turned out to be a mistake.
This is the original implementation, done while the ideas were being
designed and experimented with back in October/November.  Most of this
work took place in the Reference Library at the Milton Keynes Public
Library, which is not a bad place to work, but not as conducive to
concentration as the Reading Room at the Edinburgh Central Library,
where some of my early prc-tools work was done.  :-)

Palmdev-prep does a lot of iterating through sets of various sorts
of directories and such.  This implementation uses for_each_foo()
functions which take process() callback functions.  This technique
works well in a language in which you have nested functions with
lexical scoping (e.g. Pascal), but sadly it becomes incomprehensible
quickly in standard C, where the processor functions need to be at
global scope, and communication variables from callers to processors
need to be global too.

Since then palmdev-prep has been completely reimplemented.  So this
checkin is for posterity only.