Now that make all and make install are no longer intertwined, we can build in %build and install in %install instead of using that all-install abomination.
We also build *under* $RPM_BUILD_DIR/%{name}-%{version} instead of
beside it, so that %clean can catch everything easily.

Unfortunately rpm's %configure macro can only do an in-source
configure, so we have to write out the configure line explicitly.
(And we want to write all that out because eg %{_infodir} gives you
(on RH 7.x) a FHS-compliant /usr/share/info as opposed to the
/usr/info you would get from configure by default, without any
--infodir option.)

This change leaves the .spec file broken because the %files stanza
has not been updated.  The next checkin will fix that.