Update to the current mainline config.guess and config.sub.  These are the pristine mainline sources!  No local patches!
In the past, mainline config.sub did not recognise "m68k-palmos",
so we patched it so as to canonicalise it as "m68k-palmos-coff"
(old-style).  Current mainline config.sub handles "m68k-palmos",
canonicalising it as "m68k-unknown-palmos" (new-style).  (Both old
and new config.sub canonicalise "m68k-palmos-coff" as itself, of
course.)

Thus if your project takes a copy of config.sub from here, it will
continue to recognise "m68k-palmos", but the actual canonicalisation
will be different, if that matters for you.

It doesn't matter for prc-tools itself; specifically:

* Current prc-tools binutils has new-style configury and uses new
  mainline config.sub with a local patch to hack the canonicalisation
  of "m68k-palmos-coff" back to "m68k-unknown-palmos";

* Current prc-tools GCC and GDB have old-style configury and use old
  mainline config.sub with a local patch to canonicalise "m68k-palmos"
  as "m68k-palmos-coff";

* The other prc-tools subdirectories use this new mainline config.sub,
  but care only that the triple is recognised; the two usual suspects
  currently get canonicalised differently, but that doesn't matter.

Our GDB patches will get new-style prc-tools configury and work
similarly to current prc-tools binutils in the near future, when we
go to 5.1 or so with the rewritten remote-palmos.c.  Our GCC patches
will get new-style prc-tools configury when we rewrite the patches
for GCC 3.x (in the somewhat less near future).