Complete rewrite of HTML documentation generation.
The GNU utilities we use haven't moved to having html / install-html
targets in their makefiles very quickly, if at all.  So instead of
locally patching in such targets and being able to get what we want
only in a fragile manner, we write our own targets here, and give them
very long fingers reaching over into the other tool directories.  This
is a hack, but it's more satisfactory than the old hack.