Make palmdev-prep work with multiple targets.
Instead of being told what the target is at configure time
(TARGET_ALIAS, for example, will soon be meaningless in the tools
subdirectory), it searches for installed targets when it is run.

This also takes care of the fact that, if we package them separately
(as we plan to), a user might install only a subset of the targets
that were configured.