Remember that ARM and especially PACE do appallingly badly when you ask them to move bytes around.  So here are better memset()/memcpy() routines that move 32 bits at a time when they can.
(These functions share no source with the previous Linux ones -- there
should have been a remove/add source control operation here.  Diffs to
the previous revision are meaningless.)

This change also fixes a long-standing bug (#230368): the previous
memset() function's signature was wrong (the source char had type char
instead of the correct int).  We've deferred that for ages because of
binary compatibility fears w.r.t. a libc GLib, but that fear is no
longer important.

(Yes, memcpy()'s parameters should be "restrict", but we'll worry
about that when it's implemented :-))