On Cygwin, open files with O_TEXT.  This isn't really the right thing to do, and the issue will become moot with a 3.x preprocessor anyway (because it handles \r itself), but I can't bear the FUD messages from the clueless people who screw up their binmode/textmode mounts anymore. Oh, and it'll help them :-).  I suppose someone should have submitted this patch long ago!
The changes are copied from Cygwin's local patches to (native) GCC.

This allows the user to get away with no mounts at all.  (Sometime it
might be worth updating the installation instructions to note this.)
The intention was that the mount point gave a Palm OS SDK installer a
place to look in the Registry to detect prc-tools and the right place
to install itself, as was almost done right in the 3.5 SDK (except that
it looked in the v1 mount table, sigh); nice theory, but the reality is
that PalmSource will never again be capable of producing an SDK that
supports prc-tools to anything like that extent :-(.  So there's no
need to keep the mount point around for the SDK's sake.

(BTW the recent libiberty/fnmatch() change to the binutils and GDB
patches wasn't needed here in the GCC patches, because fnmatch() is
not used anywhere in GCC.)