Duh -- of course we need curname.  The original string might stay alive (so we can store a pointer to it instead of copying the text), but it won't be accessible via tree->dirstrack once the scanner starts pushing more fresh directories on the top of the stack!
So bring back curname, and don't look at dirstack within the
readdir() loop.

The other problem with the two "readdir is finished, look at top of
stack" handlers is it's not clear which one should be closing the
finished DIR -- in fact, either one could be required to, depending
on whether any more fresh directories got pushed.  The real answer is
to close it as soon as readdir() runs out, which we've recoded to do.

Hopefully there's no new bugs this time :-).  Must Be More Careful.