Overhaul the generic stdlib.h, and add a m68k-specific one that adds to it.  Unlike the previous stdlib.h which provided declarations for just about everything that should be in stdlib.h even if there were no implementations backing them up in libc.a, we're only adding things to this stdlib.h as they are implemented (preferably for both targets, but possibly only for one, usually m68k).
The "unlicensing" text at the top is my original draft; it may need
tweaking to take into account the comments from the prc-tools-devel
discussion.

Now that we have a m68k-specific stdlib.h, the extern inline versions
of malloc() and free() that call through to MemPtrNew/MemPtrFree are
back.

The various other declarations will be returning as they are fully
implemented.  The extern inline m68k-asm optimized div() functions,
for example, will be returning to the m68k-specific header.