Add the various *abs() and *div() functions to libc and stdlib.h.
The m68k-specific m68k/stdlib.h supplies the inline asm div()
that has been in prc-tools' stdlib.h for a long time, but with
a couple of fixes so that now it has a chance of working.  The
fact that nobody's complained possibly suggests something about
how widely used these functions are :-).