A new implementation of the <ctype.h> functions, tailored for Palm OS and for the vagaries of the Palm OS character set.
The old version's ctype.h defined a _SP that conflicted with the SDK;
this version's ctype.h fixes that bug (at last!) and chooses not to
define names for such flag masks *at all*, so that that sort of thing
won't happen again.

The ARM side (GCC 3 + __CHAR_UNSIGNED__) has not yet been tested,
and there may also be a few mistakes in the _Ctype[] table.  In
particular, some of the characters that are currently "nothing"
may have become new (probably punctuation) characters in Palm OS 5;
unfortunately AsciiChart does not run on the Simulator :-) and this
has not been checked yet.

This implementation disagrees in a few ways with the native Palm OS
functions:

 * it considers that only the characters that actually have some
   "control" characteristics should be control characters, rather
   than the native functions' "*all* non-printable characters are
   control characters";

 * it considers the two ordinal indicators to be caseless letters;

 * (at present) it considers '\x83' ("small f hook"; florin?) to
   be punctuation rather than a letter.

Fixed bug 413271