Cast the index of _Ctype to int to avoid "array subscript is of type char" warnings.  This warning is telling us that unadorned char might be signed; indeed on m68k it is, and we've taken steps to deal with that.
The other way to fix this is to admit that there's really no such
thing as EOF on m68k Palm OS, and cast to unsigned char instead.
Because the indexing addressing mode used on m68k requires this index
to end up 16 bits wide anyway, this wouldn't affect performance: we
have to sign extend the char to 16 bits anyway, so just clearing the
upper byte instead would take the same amount of time and allow us
to have a 256 byte table instead of a 384 byte one.

But I'm not ready to admit that yet :-).