Remove the old Linux-derived string.h from libc/include in favour of a new one in include.
The old header declared: strtol() and strtoul() (wrongly), which was
wrong so has been removed; something called intodec(), which wasn't
even defined anywhere so I don't know what it was and it has been
removed; and the non-standard functions memscan() and strnlen(),
which are non-standard and I suspect not used much so have been
removed (as have their old Linux-derived implementations).

If there's a problem, we could perhaps reimplement those last two,
but I don't expect that to happen.

Some of the functions (memchr(), strrchr()) declared in the new
header were missing from the old library; we will be adding new
implementations soon.  We will also be replacing the old, frequently
buggy, Linux-derived implementations with new public domain
implementations.  When this process is complete, we'll add the
usual "public domain" boilerplate to the top of the new string.h.