Collect the new memcpy() and memset() implementations into a single source file, memstring.c (to which we'll be adding the other mem* functions), and use the usual -DLmemcpy trick in the makefile.
This source code arrangement is less efficient because you have to
pass all that text through the preprocessor several times, but the
impact is trivial and IMHO outweighed by the advantages of having
all the related source code together and of not drowning in masses
of tiny source files.

Removed the quotations from the Standard describing these functions,
because we're not going to do that for every function, and because
all interested parties either know well enough what these functions
do or probably have a copy of the Standard anyway.