Rearrange the extra directories for a multiple targets build: keep the multi-target bfd (of course) and rename it to "multi-bfd", but use one of the existing libiberties instead of building yet another copy of it.
In tools/Makefile.in, specify the BFD library via an explicit path
instead of via -lbfd, so that we can be totally sure we get the one
we want and won't, for example, get a native one if our one failed
to build for some reason.  (I *knew* there was a reason for using
an explicit path instead of -lfoo!  I think the reason in the other
direction was the thought that -lfoo could possibly helpfully match
more than just libfoo.a, e.g. if libtool is in use over there, but
that's really a forlorn hope.)

This is also framework for getting libintl.a into some of the link
commands in the tools directory, so that --enable-nls etc will work
properly when it needs an explicit -lintl (which it doesn't on
Linux because the intl functions are in glibc).

There was no way I was going to build an extra libintl, so this was
a good time to get rid of the extra libiberty too.  A further change
is needed in configure.in to actually set @INTLLIBS@.

(Even though we don't use any intl functions ourselves (yet :-)),
the executables that use BFD can need to link with it, because BFD is
itself internationalised.  So we don't need to bother making headers
available, but we do need to supply appropriate link libraries.)