Set @INTLLIBS@ (more or less) appropriately.
The programs in the tools directory that link with BFD may also need
to link with some libintl, because the BFD routines use libintl.  So,
rather than adding the standard configury, we've added just enough to
find a library to link with; we (currently) don't need to worry about
headers.  (The standard macros for this would also have trouble with
our finding a libintl down in our binutils instead of adding one at
the top level.)

Tested on Cygwin, where this change makes an --enable-nls build
actually build.  However it choose the system -lintl for tools,
while the binutils directory chooses to use its included libintl.
(This might be problematic because it adds a shared library
dependency.)  Our libintl test therefore differs from that used
by binutils, GCC, etc, and may need further tweaking.

(My guess about where in configure.in I was going to want to put
this turned out to be wrong :-))