Define __INT_MAX__ correctly for our -mshort default.  For years, <limits.h> has been broken by this; I'm finally fixing it now because I'm about to add a <stdint.h> which would be broken too :-).
The proper way to do this would be similar to what is done on MIPS:
Add a %(cpp_int_max) spec in m68k-none.h, and have it test either
%{mshort} or %{!mnoshort}, depending on what TARGET_DEFAULT says
is the default for MASK_SHORT.  However, TARGET_DEFAULT is set too
late (especially in GCC 3.2) for a #if in m68k-none.h to act upon
it, so what you really have to do is put the %{mshort} version in
m68k-none.h with a note that targets for which MASK_SHORT is on by
default also need to override CPP_INT_MAX_SPEC to use %{!mnoshort}.

However the only reason to do it properly is so that it would be
worthy of submission back to mainline GCC; but GCC 3.3 already fixes
this in all cases on all targets, by having the compiler always
predefine __INT_MAX__ and friends using its generic knowledge of
what the corresponding types really are.  (And if it wasn't for that,
it would also be better to use 3.3's TARGET_CPU_CPP_BUILTINS instead.)

Thus, since the "proper" fix touches m68kpalmos.h anyway, the simplest
thing to do in GCC 2.95.3 is just the single m68kpalmos.h addition
here.  This would also be the right fix for GCC 3.2.

Fixed bug 212925 [the remaining INT_MAX aspects]

(Also simplified the use of spaces in CPP_SUBTARGET_SPEC.  Because
contiguous spaces in specs are collapsed to a single space anyway,
there's no point in being obscure just to ensure that you generate
only a single space.)