Add <stdint.h>, for both m68k and ARM.  At the moment, all the [u]int_fastN_t types are just the corresponding [u]intN_t types, except [u]int_fast16_t on ARM, which is [unsigned] int, i.e., [u]int32_t.  Possibly [u]int8_t on ARM could get similar treatment, but we'll think about that another day.
This has not been as heavily tested as I would like: in fact,
the only testing has been that the types have the right sizes.
The limit and constant macros have not yet been tested at all.
But this change has been lurking in my local patches for far too
long already :-).