Generate man pages for build-prc and palmdev-prep from the info documentation, using Pod::Man and binutils' texi2pod.pl.
The version.texi.in change is a delicious hack :-).

The original plan was to use pod2man instead of the Perl script
inline in the makefile, but this way it was easier to extract
the information from the version.texi output, and being able to
set $name explicitly allows us to use a nice tidy pipeline instead
of explicit temporary files.  Sadly, texi2pod.pl and pod2man don't
seem to report errors via their return statuses very reliably, so
we don't bother checking them; be sure to check the output.

The configure.in change is mostly reindenting: like Pod::Man,
we only actually need to check texi2html and prc_tools_version
in an --enable-generic build; and we don't need to check whether
prc-tools.spec exists because we've already checked that with
AC_INIT.

We'll need to add the manpages to prc-tools.build as well as
prc-tools.spec.