Update the Make HTML documentation build to Make 3.80.
Also rearrange the Source: lines in the spec file to be a bit more
logical: put make absolutely last, and order the others more logically
and also alphabetically -- i.e. put GDB last.

Part of the reason for doing that is that this new Make is using
new autotools.  The new autoconf puts "" for $ac_cv_exeext instead
of "no", and when that gets into config.cache it affects the rest
of the build.  So there was a hope that this rearrangement would
cause Make to be configured last and not pollute the cache, but of
course because of --enable-targets, make is the only subdir that
gets configured in the first configure, so it always gets configured
first!  Happily, the configure script appears to be immune to any
effects from this change in $ac_cv_exeext.  So Never Mind.