Recent GCC releases no longer listen to --disable-cpp, so to get rid of 3.x $target-cpp executables we need to delete them ourselves after "make install" (or hack further into the GCC makefile).
And it turns out that all this disabling installation of various
specific bits and pieces is really merely a question of packaging
policy; not even everyone building prc-tools needs to agree with
all of these decisions.

Thus their proper place is really in our package scripts rather
than as makefile patches.  After all, such makefile patches need
to be maintained locally and would never be appropriate for sending
upstream, so they're really not a win over deleting things later.

So this change also undoes the suppression of gcov and gccbug in
the GCC 3.x makefile in favour of deleting them afterwards.

The package script is now responsible for removing:
  bindir: preprocessor drivers, extra gcc drivers, gcov, gccbug;
  tooldirs: extra copies of the g++ driver;
  gcc-lib dirs: the install-tools hierarchy.

(I did actually have makefile patches to disable install-tools etc
ready to check in, but this is a better---albeit low-tech---approach.)