Disable GCC Makefile rules for parser generation via Bison.
There are a lot of assumptions about the Bison-generated parser code
baked into the GCC codebase, which leads to lots of errors when building
with newer versions of Bison. Since the intention of this repo is to
"freeze" the toolchain as-is, this change simply freezes the
Bison-generated parser code (c-parse.c etc) by disabling the Makefile
rules to regenerate them at build time.